Support ARM architecture
When converting between slices of bytes and uint64, we can use
unsafe casting on x86-64 architecture because the byte alignment
is guaranteed. However, on ARM architecture there is a possibility
this will cause issues. Testing on an ARM VM didn't cause problems
but it is best to provide an option without unsafe casting. This
commit adds build directives where necessary to support the two
versions. Where needed, the shared functionality remains in the
original filename (bit.go). Functions which rely on unsafe casting
are in a file appended with _amd64.go (bits_amdt64.go). Support
without unsafe casting is in a file appended with _gen.go
(bits_gen.go). The version without unsafe casting is not as
performant.
