+++ internal/util/bitvect.go
@@ -3,8 +3,6 @@ package util
 import (
 	"runtime"
 	"sync"
-
-	"github.com/alecthomas/unsafeslice"
 )
 
 const bitVectWidth = 512
@@ -13,40 +11,6 @@ const bitVectWidth = 512
 // uint64 elements.
 type BitVect struct {
 	set [bitVectWidth * 8]uint64
-}
-
-// unravelTall populates a BitVect from a 2D matrix of bytes. The matrix
-// must have 64 columns and a multiple of 512 rows. idx is the block target.
-// Only tested on x86-64.
-func (b *BitVect) unravelTall(matrix [][]byte, idx int) {
-	for i := 0; i < bitVectWidth; i++ {
-		copy(b.set[(i)*8:(i+1)*8], unsafeslice.Uint64SliceFromByteSlice(matrix[(bitVectWidth*idx)+i]))
-	}
-}
-
-// unravelWide populates a BitVect from a 2D matrix of bytes. The matrix
-// must have a multiple of 64 columns and 512 rows. idx is the block target.
-// Only tested on x86-64.
-func (b *BitVect) unravelWide(matrix [][]byte, idx int) {
-	for i := 0; i < bitVectWidth; i++ {
-		copy(b.set[i*8:(i+1)*8], unsafeslice.Uint64SliceFromByteSlice(matrix[i][idx*64:(64*idx)+64]))
-	}
-}
-
-// ravelToTall reconstructs a subsection of a tall (mx64) matrix from a BitVect.
-// Only tested on x86-64.
-func (b *BitVect) ravelToTall(matrix [][]byte, idx int) {
-	for i := 0; i < bitVectWidth; i++ {
-		copy(matrix[(idx*bitVectWidth)+i][:], unsafeslice.ByteSliceFromUint64Slice(b.set[i*8:(i+1)*8]))
-	}
-}
-
-// ravelToWide reconstructs a subsection of a wide (512xn) matrix from a BitVect.
-// Only tested on x86-64.
-func (b *BitVect) ravelToWide(matrix [][]byte, idx int) {
-	for i := 0; i < bitVectWidth; i++ {
-		copy(matrix[i][idx*64:(idx+1)*64], unsafeslice.ByteSliceFromUint64Slice(b.set[(i*8):(i+1)*8]))
-	}
 }
 
 // ConcurrentTransposeTall tranposes a tall (64 column) matrix. If the input
